+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:places/domain/filter.dart';
-import 'package:places/domain/sight.dart';
+import 'package:places/model/filter.dart';
+import 'package:places/model/sight.dart';
 import 'package:places/search_history_state.dart';
 import 'package:places/theme_state.dart';
 import 'package:places/ui/res/app_strings.dart';
