+++ lib/model/geo_point.dart
