+++ lib/model/sight_type.dart
