+++ lib/ui/screen/add_sight_screen/widget/sight_photo_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:places/domain/sight_photo.dart';
+import 'package:places/model/sight_photo.dart';
 import 'package:places/ui/res/app_colors.dart';
 import 'package:places/ui/res/svg_icons/svg_icon.dart';
 import 'package:places/ui/res/svg_icons/svg_icons.dart';
