+++ lib/ui/screen/sight_details_screen/sight_details_bottomsheet.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:places/domain/sight.dart';
-import 'package:places/domain/sight_photo.dart';
 import 'package:places/mocks.dart';
+import 'package:places/model/sight.dart';
+import 'package:places/model/sight_photo.dart';
 import 'package:places/ui/res/app_strings.dart';
 import 'package:places/ui/res/app_text_styles.dart';
 import 'package:places/ui/res/svg_icons/svg_icon.dart';
