+++ lib/ui/screen/sight_details_screen/sight_details_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:places/domain/sight.dart';
-import 'package:places/domain/sight_photo.dart';
 import 'package:places/mocks.dart';
+import 'package:places/model/sight.dart';
+import 'package:places/model/sight_photo.dart';
 import 'package:places/ui/res/app_strings.dart';
 import 'package:places/ui/res/app_text_styles.dart';
 import 'package:places/ui/res/svg_icons/svg_icons.dart';
