+++ lib/ui/screen/sight_list_screen/sight_list_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:places/config.dart';
-import 'package:places/domain/filter.dart';
-import 'package:places/domain/sight.dart';
 import 'package:places/filter_utils.dart';
 import 'package:places/mocks.dart';
+import 'package:places/model/filter.dart';
+import 'package:places/model/sight.dart';
 import 'package:places/ui/res/app_strings.dart';
 import 'package:places/ui/res/app_text_styles.dart';
 import 'package:places/ui/res/svg_icons/svg_icons.dart';
